Split util functions into their own file and add missing webhook properties
